Refactor most of the provider and provider builder into a common base class. Sad panda for using abstract classes.
